Somehow Managed to implemet RiverPod for now the catalogue works fine , but lost the functionality of routing
