+++ lib/Screens/RegisterPage.dart
@@ -1,32 +1,38 @@
-
-import 'package:faculty_review/SearchTeachers.dart';
+import 'package:faculty_review/Screens/LoginPage.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'RegisterPage.dart';
-import 'constants.dart';
-class LoginPage extends StatefulWidget {
-  const LoginPage({super.key});
+import '../constants.dart';
+class RegisterPage extends StatefulWidget {
+  const RegisterPage({super.key});
 
   @override
-  _LoginPageState createState() => _LoginPageState();
+  State<RegisterPage> createState() => _RegisterPageState();
 }
 
-class _LoginPageState extends State<LoginPage> {
+class _RegisterPageState extends State<RegisterPage> {
+
   bool _obscureText = true;
 
   void _togglePasswordVisibility() {
     setState(() {
       _obscureText = !_obscureText;
     });
+
+
   }
+
 
   @override
   Widget build(BuildContext context) {
-    final deviceWidth = MediaQuery.of(context).size.width;
-    final maxWidth = deviceWidth > 600 ? 600.0 : deviceWidth * 0.95;
-
+    final deviceWidth = MediaQuery
+        .of(context)
+        .size
+        .width;
+    final maxWidth = deviceWidth > 600 ? 600.0 : deviceWidth *
+        0.95; // Define a maximum width for the content
     return Scaffold(
-      appBar: Constants.constantAppBar,
+
+      appBar:  Constants.constantAppBar,
       body: Column(
         children: [
           Expanded(
@@ -58,15 +64,19 @@ class _LoginPageState extends State<LoginPage> {
                           decoration: const InputDecoration(
                             hintText: "Erp or Email",
                             focusedBorder: OutlineInputBorder(
-                              borderSide: BorderSide(color: brownColor, width: 5,),
-                              borderRadius: BorderRadius.all(Radius.circular(10)),
+                              borderSide: BorderSide(
+                                color: Color(0xff700f1a), width: 5,),
+                              borderRadius: BorderRadius.all(
+                                  Radius.circular(10)),
                             ),
                           ),
                           style: TextStyle(
                             color: Colors.black,
                             fontSize: 15,
                             fontWeight: FontWeight.bold,
-                            fontFamily: GoogleFonts.spaceMono().fontFamily,
+                            fontFamily: GoogleFonts
+                                .spaceMono()
+                                .fontFamily,
                           ),
                           textAlign: TextAlign.center,
                         ),
@@ -88,11 +98,15 @@ class _LoginPageState extends State<LoginPage> {
                           decoration: InputDecoration(
                             hintText: "Enter your Password",
                             focusedBorder: const OutlineInputBorder(
-                              borderSide: BorderSide(color: brownColor, width: 5,),
-                              borderRadius: BorderRadius.all(Radius.circular(10)),
+                              borderSide: BorderSide(
+                                color: Color(0xff700f1a), width: 5,),
+                              borderRadius: BorderRadius.all(
+                                  Radius.circular(10)),
                             ),
                             suffixIcon: IconButton(
-                              icon: Icon(_obscureText ? Icons.visibility_off : Icons.visibility),
+                              icon: Icon(
+                                  _obscureText ? Icons.visibility_off : Icons
+                                      .visibility),
                               onPressed: _togglePasswordVisibility,
                             ),
                           ),
@@ -100,7 +114,49 @@ class _LoginPageState extends State<LoginPage> {
                             color: Colors.black,
                             fontSize: 15,
                             fontWeight: FontWeight.bold,
-                            fontFamily: GoogleFonts.spaceMono().fontFamily,
+                            fontFamily: GoogleFonts
+                                .spaceMono()
+                                .fontFamily,
+                          ),
+                          textAlign: TextAlign.center,
+                        ),
+                      ),
+                    ),
+                    Padding(
+                      padding: const EdgeInsets.all(10),
+                      child: Container(
+                        decoration: BoxDecoration(
+                          borderRadius: BorderRadius.circular(10),
+                          shape: BoxShape.rectangle,
+                          border: Border.all(
+                            color: Colors.black,
+                            width: 1,
+                          ),
+                        ),
+                        child: TextFormField(
+                          obscureText: _obscureText,
+                          decoration: InputDecoration(
+                            hintText: "Confirm your Password",
+                            focusedBorder: const OutlineInputBorder(
+                              borderSide: BorderSide(
+                                color: Color(0xff700f1a), width: 5,),
+                              borderRadius: BorderRadius.all(
+                                  Radius.circular(10)),
+                            ),
+                            suffixIcon: IconButton(
+                              icon: Icon(
+                                  _obscureText ? Icons.visibility_off : Icons
+                                      .visibility),
+                              onPressed: _togglePasswordVisibility,
+                            ),
+                          ),
+                          style: TextStyle(
+                            color: Colors.black,
+                            fontSize: 15,
+                            fontWeight: FontWeight.bold,
+                            fontFamily: GoogleFonts
+                                .spaceMono()
+                                .fontFamily,
                           ),
                           textAlign: TextAlign.center,
                         ),
@@ -112,12 +168,7 @@ class _LoginPageState extends State<LoginPage> {
                           child: Padding(
                             padding: const EdgeInsets.all(10),
                             child: ElevatedButton(
-                              onPressed: () {
-                                Navigator.push(
-                                  context,
-                                  MaterialPageRoute(builder: (context) => const SearchTeachers()),
-                                );
-                              },
+                              onPressed: () {},
                               style: ElevatedButton.styleFrom(
                                 backgroundColor: const Color(0xff700f1a),
                                 padding: const EdgeInsets.all(10),
@@ -126,7 +177,7 @@ class _LoginPageState extends State<LoginPage> {
                                 ),
                               ),
                               child: const Text(
-                                'Login',
+                                'Register',
                                 style: TextStyle(
                                   color: Colors.white,
                                   fontSize: 15,
@@ -141,17 +192,16 @@ class _LoginPageState extends State<LoginPage> {
                     TextButton (onPressed: () {
                       Navigator.push(
                         context,
-                        MaterialPageRoute(builder: (context) => RegisterPage()),
+                        MaterialPageRoute(builder: (context) => const LoginPage()),
                       );
                     },
 
                       child: const Text
-                        ('Register?', style: TextStyle(
+                        ('Already Have an account? Login.', style: TextStyle(
                         color: Colors.black,
                         fontSize: 15,
                         fontWeight: FontWeight.bold,
                       ),),),
-
                   ],
                 ),
               ),
@@ -164,7 +214,9 @@ class _LoginPageState extends State<LoginPage> {
               child: Text(
                 'Dev by: Zohaib Mughal',
                 style: TextStyle(
-                  fontFamily: GoogleFonts.roboto().fontFamily,
+                  fontFamily: GoogleFonts
+                      .roboto()
+                      .fontFamily,
                   fontSize: 15,
                   color: Colors.black,
                 ),
@@ -176,4 +228,5 @@ class _LoginPageState extends State<LoginPage> {
       backgroundColor: const Color(0xffffffff),
     );
   }
+
 }
